Refactoring diffs
