+++ src/Cody.VisualStudio.Completions/Cody.VisualStudio.Completions.csproj
@@ -47,9 +47,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Completions\Difference.cs" />
-    <Compile Include="Completions\StringDifference.cs" />
     <Compile Include="Completions\StringDifference2.cs" />
-    <Compile Include="Completions\VsDifference.cs" />
     <Compile Include="LoggerFactory.cs" />
     <Compile Include="Completions\CodyProposalCollection.cs" />
     <Compile Include="Completions\CodyProposalManager.cs" />
